The animations moved into separate files.
