+++ drape_frontend/animation/interpolations.hpp
@@ -9,16 +9,6 @@ namespace df
 
 double InterpolateDouble(double startV, double endV, double t);
 m2::PointD InterpolatePoint(m2::PointD const & startPt, m2::PointD const & endPt, double t);
-
-class InerpolateAngle
-{
-public:
-  InerpolateAngle(double startAngle, double endAngle);
-  double Interpolate(double t) const;
-
-private:
-  double m_startAngle;
-  double m_delta;
-};
+double InterpolateAngle(double startAngle, double endAngle, double t);
 
 } // namespace df
