+++ drape_frontend/user_event_stream.cpp
@@ -1,4 +1,9 @@
 #include "drape_frontend/user_event_stream.hpp"
+#include "drape_frontend/animation/linear_animation.hpp"
+#include "drape_frontend/animation/scale_animation.hpp"
+#include "drape_frontend/animation/follow_animation.hpp"
+#include "drape_frontend/animation/perspective_animation.hpp"
+#include "drape_frontend/animation/sequence_animation.hpp"
 #include "drape_frontend/animation_constants.hpp"
 #include "drape_frontend/animation_system.hpp"
 #include "drape_frontend/animation_utils.hpp"
